fix: 修复 review 发现的 D1/D2/D3 缺陷 + 补充 D6 回归测试
- D1: fuse_direction 未过滤无效超声读数(真机超时 -1.0cm 被 L0 误判为超近障碍 → 假性急停)
      现在仅 valid 读数参与融合, 否则兜底 4.5m
- D2: layer_fusion/calc_confidence 的超声有效区间统一为 [0.02m, 4.5m),
      L0 盲区分支加下限, 负距离不再被当成有效值参与加权
- D3: TSL2591 I2C 读取失败不再返回 0(被当室内→深度相机最高权重),
      改为返回 -1 触发 determine_environment 回退 OUTDOOR(超声波主导, 安全侧)
- D6: 新增 test_invalid_ultrasonic_no_false_critical 回归测试,
      断言无效超声输入不产生负距离/CRITICAL

验证: WSL gcc 13.3 编译通过, ctest 100% (test_fusion 210 项断言全绿)

+++ sensor_ir.h
@@ -44,7 +44,7 @@ public:
     explicit InfraRedSensor(bool use_simulated = true);
     ~InfraRedSensor();
 
-    /** 读取归一化环境红外强度: 0.0(暗) ~ 1.0(强红外) */
+    /** 读取归一化环境红外强度: 0.0(暗) ~ 1.0(强红外); 返回 -1.0 表示读取失败 */
     double read_normalized_light();
 
 private:
@@ -61,7 +61,7 @@ private:
     bool i2c_open();
     void i2c_close();
     bool tsl2591_init();
-    uint16_t tsl2591_read_ir(); // 读 CHAN1 (infrared) 通道
+    int tsl2591_read_ir();      // 读 CHAN1 (infrared) 通道; 返回 -1 表示失败
     bool i2c_write_reg(uint8_t reg, uint8_t value);
     bool i2c_read_reg(uint8_t reg, uint8_t* out, uint8_t len);
 #endif
